The GaspFixer is actually pretty cool but needs to be in gftools.fix
